Initial breakup complete
